+++ src/main/java/by/toukach/cleverbank/view/SpendAccountViewChain.java
@@ -20,7 +20,7 @@ public class SpendAccountViewChain extends TransactionViewChain {
     System.out.println("Введите сумму");
     Long accountId = getAccountDto().getId();
     Scanner scanner = getScanner();
-    long answer = scanner.nextLong();
+    double answer = scanner.nextDouble();
     scanner.nextLine();
 
     TransactionDto transactionDto = TransactionDto.builder()
